+++ src/vs/editor/contrib/hover/browser/contentHoverController2.ts
