+++ src/vs/workbench/contrib/chat/browser/codeBlockPart.ts
@@ -64,7 +64,7 @@ import { IMarkdownVulnerability } from '../common/annotations';
 import { ResourceLabel } from 'vs/workbench/browser/labels';
 import { FileKind } from 'vs/platform/files/common/files';
 import { ResourceContextKey } from 'vs/workbench/common/contextkeys';
-import { ContentHoverController } from 'vs/editor/contrib/hover/browser/contentHoverController';
+import { ContentHoverController } from 'vs/editor/contrib/hover/browser/contentHoverController2';
 import { MarginHoverController } from 'vs/editor/contrib/hover/browser/marginHoverController';
 
 const $ = dom.$;
